+++ src/static_file_data.rs
@@ -21,6 +21,9 @@ pub const ALL_LOC_CSV_NAME: &str = "Loc.csv";
 /// The file name of the text file with only the translation keys.
 pub const LOC_REF_NAME: &str = "loc_keys.txt";
 
+/// The file name of the text file with the music file names.
+pub const MUSIC_REF_NAME: &str = "music.txt";
+
 /// The file name of the sound reference file.
 pub const SOUND_REF_NAME: &str = "sounds.slop";